Revert "[BE]읽음처리 API"
