+++ backend-api/src/main/java/org/wildflowergardening/backend/api/wildflowergardening/application/HomelessAppService.java
@@ -54,8 +54,6 @@ public class HomelessAppService {
     private final DailyOutingCountsService dailyOutingCountsService;
     private final DailySleepoverCountsService dailySleepoverCountsService;
     private final DailyEmergencyCountsService dailyEmergencyCountsService;
-    private final NoticeTargetService noticeTargetService;
-    private final NoticeService noticeService;
 
     public List<HomelessTermsResponse> getAllTerms() {
         return homelessTermsService.findAll(LocalDate.now()).stream()
@@ -304,7 +302,7 @@ public class HomelessAppService {
     }
 
     @Transactional
-    public void saveEmergencyLog(Long homelessId, Long shelterId, EmergencyRequest request) {
+    public void saveEmergencyLog(long homelessId, long shelterId, EmergencyRequest request) {
 
         Homeless homeless = homelessQueryService.getOneById(homelessId)
                 .orElseThrow(() -> new IllegalArgumentException("해당 사용자를 찾을 수 없습니다."));
@@ -323,7 +321,7 @@ public class HomelessAppService {
 
     }
 
-    public String getStatusLocationByHomelessId(Long homelessId, Long shelterId) {
+    public String getStatusLocationByHomelessId(long homelessId, long shelterId) {
         LocationTracking locationTracking = locationTrackingService.getLocationByHomelessId(homelessId, shelterId);
 
         InOutStatus status = locationTracking.getInOutStatus();
@@ -332,10 +330,5 @@ public class HomelessAppService {
         } else {
             return "OUT_SHELTER";
         }
-    }
-
-    @Transactional
-    public void updateNoticeReadStatus(Long homelessId, Long noticeId, boolean status) {
-        noticeTargetService.updateReadStatus(noticeId, homelessId, status);
     }
 }
