--- backend-api/src/main/java/org/wildflowergardening/backend/api/wildflowergardening/application/dto/response/HomelessNoticeResponse.java
@@ -1,10 +1,0 @@
-package org.wildflowergardening.backend.api.wildflowergardening.application.dto.response;
-
-import io.swagger.v3.oas.annotations.media.Schema;
-import lombok.Builder;
-import lombok.Getter;
-
-@Getter
-@Builder
-public class HomelessNoticeResponse {
-}
