+++ backend-core/src/main/java/org/wildflowergardening/backend/core/wildflowergardening/domain/NoticeTarget.java
@@ -33,10 +33,10 @@ public class NoticeTarget {
     @Comment("노숙인 id")
     private Long homelessId;
 
-    @Column(name = "read_status", nullable = false)
+    @Column(name = "is_read", nullable = false)
     @Builder.Default
     @Setter
-    private boolean readStatus = false;
+    private boolean isRead = false;
 
     @Column(name = "read_at")
     @Comment("읽은 시간")
