+++ backend-core/src/main/java/org/wildflowergardening/backend/core/wildflowergardening/domain/NoticeTargetRepository.java
@@ -3,8 +3,5 @@ package org.wildflowergardening.backend.core.wildflowergardening.domain;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import java.util.List;
-
 public interface NoticeTargetRepository extends JpaRepository<NoticeTarget, Long> {
-    List<NoticeTarget> getNoticeTargetByNoticeIdAndHomelessId(Long noticeId, Long homelessId);
 }
